External files links fixed
Many external file links fixed. This will significantly improve
performance and prevents double updates to both files if there is an
update to the interface. Interface will be the same for both pages,
while the content can be different. Reloading of the same files are
minimized and the browser may load these files from cache.